util/opentelemetry-util-genai/example/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-util-genai
opentelemetry-exporter-otlp-proto-grpc
python-dotenv
